Avoid cloning tensors in non-inference mode.
